requirements.txt
requests==2.28.1
python-dotenv==0.21.*
python-telegram-bot==13.*
redis==4.3.4